requirements.txt
streamlit==1.32.2

langchain==0.1.20
langgraph==0.0.40

langchain-google-genai==1.0.3
google-generativeai==0.5.4

python-dotenv==1.0.1

pydantic==2.6.4
typing-extensions==4.10.0